save and load test passing

+++ app/src/main/java/de/htw/ai/ema/persistence/dao/DAO.java
@@ -18,7 +18,7 @@ public interface DAO {
      * This method loads a game from storage.
      * @param id
      */
-    public Game loadGame(int id);
+    public Game loadGame(long id);
 
     /**
      * This method loads all games from storage.
